refactor: extract ExpenseRow from SwipeableExpenseRow and remove enableSwipe prop
Extract row display logic into ExpenseRow, refactor SwipeableExpenseRow
to wrap it, and update HomeScreen to use ExpenseRow directly.
