+++ src/components/record/SwipeableExpenseRow.tsx
@@ -5,24 +5,14 @@ import { YStack, XStack } from '@tamagui/stacks';
 import { SizableText } from '@tamagui/text';
 import { FontAwesome6 } from '@expo/vector-icons';
 import { fontSizes, wanchoColors } from '../../../tamagui.config';
-import { CATEGORIES } from '@/constants/categories';
 import { Expense } from '@/types/expense';
+import { ExpenseRow } from './ExpenseRow';
 
 type SwipeableRowProps = {
   item: Expense;
   onEdit?: (item: Expense) => void;
   onDelete?: (item: Expense) => void;
   showDate: boolean;
-  enableSwipe?: boolean;
-}
-
-function formatAmount(amount: number): string {
-  return `¥${amount.toLocaleString()}`;
-}
-
-function formatDate(dateStr: string): string {
-  const [, month, day] = dateStr.split('-');
-  return `${Number(month)}/${Number(day)}`;
 }
 
 // 各費用リストの1行分のコンポーネント
@@ -31,9 +21,7 @@ export const SwipeableExpenseRow = ({
   onEdit,
   onDelete,
   showDate,
-  enableSwipe = true,
 } : SwipeableRowProps) => {
-  const category = CATEGORIES.find((c) => c.id === item.categoryId);
   const ref = useRef<SwipeableMethods | null>(null);
 
   const handleEdit = (item: Expense) => {
@@ -81,63 +69,16 @@ export const SwipeableExpenseRow = ({
     )
   };
 
-  const rowContent = (
-    <XStack
-      paddingVertical={12}
-      paddingHorizontal={16}
-      backgroundColor="$white"
-      borderBottomWidth={1}
-      borderBottomColor="$lightGray"
-      alignItems="center"
-      gap={12}
-    >
-      <YStack
-        width={36}
-        height={36}
-        borderRadius={18}
-        borderColor={category?.bgColor ?? wanchoColors.sage}
-        borderWidth={1}
-        alignItems="center"
-        justifyContent="center"
-      >
-        <FontAwesome6
-          name={(category?.icon ?? 'ellipsis') as any}
-          size={18}
-          color={category?.bgColor ?? wanchoColors.greige}
-        />
-      </YStack>
-      <YStack flex={1}>
-        <SizableText fontSize={fontSizes.body} fontWeight="bold" color="$charcoal">
-          {category?.name ?? 'その他'}
-        </SizableText>
-        <XStack gap={4}>
-          {showDate && (
-            <SizableText fontSize={fontSizes.caption} color="$greige">
-              {formatDate(item.date)}
-            </SizableText>
-          )}
-          {item.memo && (
-            <SizableText fontSize={fontSizes.caption} color="$greige">
-              {item.memo}
-            </SizableText>
-          )}
-        </XStack>
-      </YStack>
-      <SizableText fontSize={fontSizes.body} fontWeight="bold" color="$charcoal">
-        {formatAmount(item.amount)}
-      </SizableText>
-    </XStack>
-  );
-
-  if (!enableSwipe) return rowContent;
-
   return (
     <ReanimatedSwipeable
       ref={ref}
       renderRightActions={() => renderRightActions(item)}
       overshootRight={false}
     >
-      {rowContent}
+      <ExpenseRow 
+        item={item}
+        showDate={showDate}
+      />
     </ReanimatedSwipeable>
   );
 };
